Refactor large app surfaces and shell compatibility
