+++ apps/admin/features/mission-control/shell/components/app-shell.ts
@@ -1,8 +1,10 @@
-export { AppShell } from "./app-shell/AppShell";
-export { GlobalSearch } from "./app-shell/GlobalSearch";
-export { MobileSidebar } from "./app-shell/MobileSidebar";
-export { NotificationsMenu } from "./app-shell/NotificationsMenu";
-export { ProfileMenu } from "./app-shell/ProfileMenu";
-export { SidebarNav } from "./app-shell/SidebarNav";
-export { TenantSwitcher } from "./app-shell/TenantSwitcher";
-export { TopBar } from "./app-shell/TopBar";
+export {
+  AppShell,
+  GlobalSearch,
+  MobileSidebar,
+  NotificationsMenu,
+  ProfileMenu,
+  SidebarNav,
+  TenantSwitcher,
+  TopBar,
+} from "../../components";
