+++ apps/admin/features/mission-control/shell/components/tiles.ts
@@ -1,5 +1,5 @@
-export { MissionControlHome } from "./tiles/MissionControlHome";
-export { QuickActionsRow } from "./tiles/QuickActionsRow";
-export { TileCard } from "./tiles/TileCard";
-export { TilePage } from "./tiles/TilePage";
-export { WorkflowsPanel } from "./tiles/WorkflowsPanel";
+export { MissionControlHome } from "../../components/tiles/mission-control-home";
+export { QuickActionsRow } from "../../components/tiles/quick-actions-row";
+export { TileCard } from "../../components/tiles/tile-card";
+export { TilePage } from "../../components/tiles/tile-page";
+export { WorkflowsPanel } from "../../components/tiles/workflows-panel";
