+++ apps/missionary/app/donors/page.tsx
@@ -4,10 +4,9 @@
 import { useMissionaryDonorRows } from "@asym/database/hooks";
 import { invalidateSupabaseTableQuery } from "@asym/database/query-keys";
 import { createBrowserClient } from "@asym/database/supabase";
-import { useAuth, useTasks } from "@asym/lib/hooks";
+import { useAuth } from "@asym/lib/hooks";
 import { motion, AnimatePresence } from "@asym/lib/motion";
 import { AddPartnerDialog } from "@asym/missionary/components/add-partner-dialog";
-import { TaskDialog } from "@asym/missionary/components/task-dialog";
 import {
   Avatar,
   AvatarFallback,
@@ -16,7 +15,6 @@ import {
 import { Badge } from "@asym/ui/components/shadcn/badge";
 import { Button } from "@asym/ui/components/shadcn/button";
 import { Card, CardContent } from "@asym/ui/components/shadcn/card";
-import { Checkbox } from "@asym/ui/components/shadcn/checkbox";
 import {
   type ColumnDef,
   DataTableColumnHeader,
@@ -79,26 +77,40 @@ import {
   Tag,
   X,
   Check,
-  CheckCircle2,
-  CreditCard,
   Building2,
   Globe,
-  DollarSign,
-  TrendingUp,
   Users,
   Star,
   Home,
-  Gift,
   Loader2,
   Repeat,
-  ListTodo,
 } from "lucide-react";
 import * as React from "react";
 import { toast } from "sonner";
 
+import { DonorTasks } from "./donor-tasks";
+import {
+  AVAILABLE_TAGS,
+  formatCurrency,
+  getActivityBg,
+  getActivityIcon,
+  getGiftTypeIcon,
+  getPaymentMethodIcon,
+  getRecurringStatusBadge,
+  getStatusBadge,
+  getStatusColor,
+  getTagLabel,
+  getTagStyle,
+} from "./donors-model";
 import { EditDonorDialog } from "./edit-donor-dialog";
 
-import type { Task } from "@asym/lib/hooks/use-tasks";
+import type {
+  Activity,
+  ActivityType,
+  Address,
+  Donor,
+  RecurringStatus,
+} from "./donors-model";
 
 import { PageHeader } from "@/components/page-header";
 
@@ -137,312 +149,6 @@ const springTransition = {
 const smoothTransition = {
   duration: 0.2,
   ease: [0.25, 0.1, 0.25, 1] as [number, number, number, number],
-};
-
-type ActivityType =
-  | "gift"
-  | "note"
-  | "call"
-  | "email"
-  | "meeting"
-  | "pledge_started"
-  | "pledge_completed";
-type GiftType =
-  | "Online"
-  | "Check"
-  | "Cash"
-  | "Bank Transfer"
-  | "Stock"
-  | "In-Kind";
-type RecurringStatus = "active" | "completed" | "paused" | "cancelled";
-
-interface Activity {
-  id: string;
-  type: ActivityType;
-  date: string;
-  title: string;
-  description?: string;
-  amount?: number;
-  status?: string;
-  gift_type?: GiftType;
-  note?: string;
-}
-
-interface RecurringDonation {
-  id: string;
-  amount: number;
-  frequency: string;
-  status: RecurringStatus;
-  start_date: string;
-  end_date?: string;
-  next_payment_date?: string;
-  total_paid: number;
-  total_expected: number;
-  payments_completed: number;
-  payments_remaining: number;
-  payment_method?: string;
-}
-
-interface Address {
-  street?: string;
-  street2?: string;
-  city?: string;
-  state?: string;
-  zip?: string;
-  country?: string;
-}
-
-interface Donor {
-  id: string;
-  name: string;
-  initials: string;
-  type: "Individual" | "Organization" | "Church";
-  status: "Active" | "Lapsed" | "New" | "At Risk";
-  total_given: number;
-  last_gift_date: string | null;
-  last_gift_amount: number | null;
-  frequency: string;
-  email: string;
-  phone: string;
-  mobile?: string;
-  work_phone?: string;
-  preferred_contact: "email" | "phone" | "text";
-  avatar_url?: string;
-  location: string;
-  address: Address;
-  work_address?: Address;
-  website?: string;
-  organization?: string;
-  title?: string;
-  joined_date: string;
-  birthday?: string;
-  anniversary?: string;
-  spouse?: string;
-  notes?: string;
-  tags: string[];
-  score: number;
-  activities: Activity[];
-  recurring_donations: RecurringDonation[];
-  has_active_pledge: boolean;
-}
-
-const AVAILABLE_TAGS = [
-  {
-    id: "major-donor",
-    label: "Major Donor",
-    color: "bg-amber-50 text-amber-700 border-amber-200",
-  },
-  {
-    id: "monthly-partner",
-    label: "Monthly Partner",
-    color: "bg-emerald-50 text-emerald-700 border-emerald-200",
-  },
-  {
-    id: "prayer-partner",
-    label: "Prayer Partner",
-    color: "bg-purple-50 text-purple-700 border-purple-200",
-  },
-  {
-    id: "church-contact",
-    label: "Church Contact",
-    color: "bg-blue-50 text-blue-700 border-blue-200",
-  },
-  {
-    id: "family",
-    label: "Family",
-    color: "bg-rose-50 text-rose-700 border-rose-200",
-  },
-  {
-    id: "friend",
-    label: "Friend",
-    color: "bg-cyan-50 text-cyan-700 border-cyan-200",
-  },
-  {
-    id: "first-time-giver",
-    label: "First-Time Giver",
-    color: "bg-indigo-50 text-indigo-700 border-indigo-200",
-  },
-  {
-    id: "legacy-giver",
-    label: "Legacy Giver",
-    color: "bg-zinc-100 text-zinc-700 border-zinc-200",
-  },
-  {
-    id: "volunteer",
-    label: "Volunteer",
-    color: "bg-orange-50 text-orange-700 border-orange-200",
-  },
-  {
-    id: "board-member",
-    label: "Board Member",
-    color: "bg-slate-100 text-slate-700 border-slate-200",
-  },
-  {
-    id: "needs-followup",
-    label: "Needs Follow-up",
-    color: "bg-red-50 text-red-700 border-red-200",
-  },
-  {
-    id: "lapsed-donor",
-    label: "Lapsed Donor",
-    color: "bg-gray-100 text-gray-600 border-gray-200",
-  },
-];
-
-const formatCurrency = (value: number | null | undefined) => {
-  if (value === null || value === undefined) return "$0";
-  return new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency: "USD",
-    minimumFractionDigits: 0,
-  }).format(value);
-};
-
-const getStatusColor = (status: string) => {
-  switch (status) {
-    case "Active":
-      return "bg-emerald-500";
-    case "Lapsed":
-      return "bg-zinc-400";
-    case "New":
-      return "bg-blue-500";
-    case "At Risk":
-      return "bg-amber-500";
-    default:
-      return "bg-zinc-400";
-  }
-};
-
-const getStatusBadge = (status: string) => {
-  const styles: Record<string, string> = {
-    Active: "bg-emerald-50 text-emerald-700 border-emerald-100",
-    Lapsed: "bg-zinc-100 text-zinc-500 border-zinc-200",
-    New: "bg-blue-50 text-blue-700 border-blue-100",
-    "At Risk": "bg-amber-50 text-amber-700 border-amber-100",
-  };
-  return (
-    <Badge
-      variant="outline"
-      className={cn(
-        "font-black border text-[9px] uppercase tracking-widest px-2 py-0.5 rounded-full",
-        styles[status] || styles["Lapsed"],
-      )}
-    >
-      {status}
-    </Badge>
-  );
-};
-
-const getRecurringStatusBadge = (status: RecurringStatus) => {
-  const styles: Record<RecurringStatus, string> = {
-    active: "bg-emerald-50 text-emerald-700 border-emerald-100",
-    completed: "bg-blue-50 text-blue-700 border-blue-100",
-    paused: "bg-amber-50 text-amber-700 border-amber-100",
-    cancelled: "bg-zinc-100 text-zinc-500 border-zinc-200",
-  };
-  return (
-    <Badge
-      variant="outline"
-      className={cn(
-        "font-black border text-[9px] uppercase tracking-widest px-2 py-0.5 rounded-full",
-        styles[status],
-      )}
-    >
-      {status}
-    </Badge>
-  );
-};
-
-const getActivityIcon = (type: ActivityType) => {
-  switch (type) {
-    case "gift":
-      return <Heart className="h-3.5 w-3.5 text-white" />;
-    case "call":
-      return <Phone className="h-3.5 w-3.5 text-white" />;
-    case "email":
-      return <Mail className="h-3.5 w-3.5 text-white" />;
-    case "note":
-      return <MessageSquare className="h-3.5 w-3.5 text-white" />;
-    case "meeting":
-      return <Briefcase className="h-3.5 w-3.5 text-white" />;
-    case "pledge_started":
-      return <TrendingUp className="h-3.5 w-3.5 text-white" />;
-    case "pledge_completed":
-      return <Check className="h-3.5 w-3.5 text-white" />;
-    default:
-      return <Clock className="h-3.5 w-3.5 text-white" />;
-  }
-};
-
-const getActivityBg = (type: ActivityType) => {
-  switch (type) {
-    case "gift":
-      return "bg-rose-500";
-    case "call":
-      return "bg-blue-500";
-    case "email":
-      return "bg-purple-500";
-    case "note":
-      return "bg-zinc-600";
-    case "meeting":
-      return "bg-emerald-500";
-    case "pledge_started":
-      return "bg-indigo-500";
-    case "pledge_completed":
-      return "bg-teal-500";
-    default:
-      return "bg-zinc-400";
-  }
-};
-
-const getGiftTypeIcon = (type: GiftType | string | undefined) => {
-  switch (type) {
-    case "Online":
-      return <CreditCard className="h-3.5 w-3.5" />;
-    case "Check":
-      return <Mail className="h-3.5 w-3.5" />;
-    case "Cash":
-      return <DollarSign className="h-3.5 w-3.5" />;
-    case "Bank Transfer":
-      return <Building2 className="h-3.5 w-3.5" />;
-    case "Stock":
-      return <TrendingUp className="h-3.5 w-3.5" />;
-    case "In-Kind":
-      return <Gift className="h-3.5 w-3.5" />;
-    default:
-      return <DollarSign className="h-3.5 w-3.5" />;
-  }
-};
-
-const getPaymentMethodIcon = (method: string | undefined) => {
-  switch (method) {
-    case "Online":
-      return <CreditCard className="h-4 w-4 text-blue-500" />;
-    case "Check":
-      return <Mail className="h-4 w-4 text-zinc-500" />;
-    case "Cash":
-      return <DollarSign className="h-4 w-4 text-emerald-500" />;
-    case "Bank Transfer":
-      return <Building2 className="h-4 w-4 text-indigo-500" />;
-    default:
-      return <CreditCard className="h-4 w-4 text-zinc-400" />;
-  }
-};
-
-const getTagStyle = (tagId: string) => {
-  const tag = AVAILABLE_TAGS.find((t) => t.id === tagId);
-  return tag?.color || "bg-zinc-100 text-zinc-600 border-zinc-200";
-};
-
-const getTagLabel = (tagId: string) => {
-  const tag = AVAILABLE_TAGS.find((t) => t.id === tagId);
-  return (
-    tag?.label ||
-    tagId
-      .split("-")
-      .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-      .join(" ")
-  );
 };
 
 function DonorListSkeleton() {
@@ -587,340 +293,6 @@ function StatCard({
 }
 
 type SortOption = "name" | "last_gift" | "total_given" | "joined_date";
-
-const TASK_TYPE_CONFIG: Record<
-  string,
-  { label: string; icon: React.ElementType; color: string; bgColor: string }
-> = {
-  call: {
-    label: "Call",
-    icon: Phone,
-    color: "text-blue-600",
-    bgColor: "bg-blue-50",
-  },
-  email: {
-    label: "Email",
-    icon: Mail,
-    color: "text-purple-600",
-    bgColor: "bg-purple-50",
-  },
-  to_do: {
-    label: "To-do",
-    icon: CheckCircle2,
-    color: "text-zinc-600",
-    bgColor: "bg-zinc-100",
-  },
-  follow_up: {
-    label: "Follow Up",
-    icon: User,
-    color: "text-orange-600",
-    bgColor: "bg-orange-50",
-  },
-  thank_you: {
-    label: "Thank You",
-    icon: Heart,
-    color: "text-rose-600",
-    bgColor: "bg-rose-50",
-  },
-  meeting: {
-    label: "Meeting",
-    icon: Users,
-    color: "text-emerald-600",
-    bgColor: "bg-emerald-50",
-  },
-};
-
-function DonorTasks({
-  donorId,
-  donorName,
-}: {
-  donorId: string;
-  donorName: string;
-}) {
-  const {
-    filteredTasks,
-    loading,
-    completeTask,
-    reopenTask,
-    deleteTask,
-    refresh,
-  } = useTasks({ donorId });
-  const [taskDialogOpen, setTaskDialogOpen] = React.useState(false);
-  const [editingTask, setEditingTask] = React.useState<Task | null>(null);
-
-  const activeTasks = filteredTasks.filter(
-    (t) => t.status !== "completed" && t.status !== "deferred",
-  );
-  const completedTasks = filteredTasks.filter((t) => t.status === "completed");
-
-  const handleComplete = async (task: Task) => {
-    if (task.status === "completed") {
-      await reopenTask(task.id);
-    } else {
-      await completeTask(task.id);
-    }
-  };
-
-  const handleEdit = (task: Task) => {
-    setEditingTask(task);
-    setTaskDialogOpen(true);
-  };
-
-  const handleTaskSuccess = () => {
-    refresh();
-    setEditingTask(null);
-    setTaskDialogOpen(false);
-  };
-
-  if (loading) {
-    return (
-      <div className="space-y-3">
-        {Array.from({ length: 3 }).map((_, i) => (
-          <div
-            key={i}
-            className="flex items-start gap-3 p-4 border rounded-xl bg-white animate-pulse"
-          >
-            <div className="h-5 w-5 rounded-md bg-zinc-200 mt-0.5" />
-            <div className="h-9 w-9 rounded-lg bg-zinc-200" />
-            <div className="flex-1 space-y-2">
-              <div className="h-4 w-3/4 bg-zinc-200 rounded" />
-              <div className="h-3 w-1/2 bg-zinc-200 rounded" />
-            </div>
-          </div>
-        ))}
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6">
-      <motion.div {...fadeInUp} className="flex items-center justify-between">
-        <div>
-          <h3 className="text-sm font-bold text-zinc-900">Tasks</h3>
-          <p className="text-xs text-zinc-500 mt-0.5">
-            Follow-ups and actions for {donorName}
-          </p>
-        </div>
-        <TaskDialog
-          task={editingTask}
-          defaultDonorId={donorId}
-          open={taskDialogOpen}
-          onOpenChange={(open) => {
-            setTaskDialogOpen(open);
-            if (!open) setEditingTask(null);
-          }}
-          onSuccess={handleTaskSuccess}
-          trigger={
-            <motion.div whileHover={{ scale: 1.02 }} whileTap={{ scale: 0.98 }}>
-              <Button size="sm" className="h-8 px-3 text-xs rounded-xl">
-                <Plus className="h-3.5 w-3.5 mr-1.5" /> Add Task
-              </Button>
-            </motion.div>
-          }
-        />
-      </motion.div>
-
-      {filteredTasks.length === 0 ? (
-        <motion.div
-          {...fadeInUp}
-          className="flex flex-col items-center justify-center py-12 text-center bg-zinc-50 rounded-2xl border border-zinc-100"
-        >
-          <div className="w-14 h-14 bg-white rounded-xl flex items-center justify-center mb-4 shadow-sm">
-            <ListTodo className="h-6 w-6 text-zinc-300" />
-          </div>
-          <p className="text-sm font-bold text-zinc-900">No tasks yet</p>
-          <p className="text-xs text-zinc-400 mt-1 max-w-[240px]">
-            Create a task to track follow-ups with this partner.
-          </p>
-        </motion.div>
-      ) : (
-        <div className="space-y-4">
-          {activeTasks.length > 0 && (
-            <div className="space-y-2">
-              <p className="text-[10px] font-black uppercase tracking-widest text-zinc-400 px-1">
-                Active ({activeTasks.length})
-              </p>
-              {activeTasks.map((task, i) => {
-                const typeConfig =
-                  TASK_TYPE_CONFIG[task.task_type] ?? TASK_TYPE_CONFIG.to_do!;
-                if (!typeConfig) return null;
-                const Icon = typeConfig.icon;
-                const isOverdue =
-                  task.due_date && new Date(task.due_date) < new Date();
-                const isDueToday =
-                  task.due_date &&
-                  new Date(task.due_date).toDateString() ===
-                    new Date().toDateString();
-
-                return (
-                  <motion.div
-                    key={task.id}
-                    initial={{ opacity: 0, y: 10 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    transition={{ delay: i * 0.03 }}
-                    className="flex items-start gap-3 p-4 border border-zinc-100 rounded-xl bg-white hover:border-zinc-200 transition-all group"
-                  >
-                    <motion.div whileTap={{ scale: 0.97 }} className="mt-0.5">
-                      <Checkbox
-                        checked={false}
-                        onCheckedChange={() => handleComplete(task)}
-                        className="h-5 w-5 rounded-md"
-                      />
-                    </motion.div>
-                    <div
-                      className={cn(
-                        "h-9 w-9 rounded-lg flex items-center justify-center shrink-0",
-                        typeConfig.bgColor,
-                        typeConfig.color,
-                      )}
-                    >
-                      <Icon className="h-4 w-4" />
-                    </div>
-                    <div className="flex-1 min-w-0">
-                      <div className="flex items-center gap-2 flex-wrap">
-                        <p className="text-sm font-bold text-zinc-900">
-                          {task.title}
-                        </p>
-                        {task.priority === "high" && (
-                          <Badge className="bg-rose-50 text-rose-600 border-0 text-[9px] font-black uppercase tracking-widest px-1.5 h-4">
-                            High
-                          </Badge>
-                        )}
-                      </div>
-                      {task.description && (
-                        <p className="text-xs text-zinc-500 mt-0.5 line-clamp-1">
-                          {task.description}
-                        </p>
-                      )}
-                      {task.due_date && (
-                        <div
-                          className={cn(
-                            "inline-flex items-center gap-1 mt-2 px-2 py-0.5 rounded-full text-[10px] font-bold uppercase tracking-wider border",
-                            isOverdue
-                              ? "bg-rose-50 text-rose-600 border-rose-100"
-                              : isDueToday
-                                ? "bg-amber-50 text-amber-600 border-amber-100"
-                                : "bg-zinc-100 text-zinc-600 border-zinc-200",
-                          )}
-                        >
-                          <Clock className="h-3 w-3" />
-                          {isOverdue
-                            ? "Overdue"
-                            : isDueToday
-                              ? "Due Today"
-                              : format(new Date(task.due_date), "MMM d")}
-                        </div>
-                      )}
-                    </div>
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          className="h-8 w-8 rounded-lg opacity-0 group-hover:opacity-100 transition-opacity"
-                        >
-                          <MoreHorizontal className="h-4 w-4" />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <DropdownMenuContent align="end" className="rounded-xl">
-                        <DropdownMenuItem
-                          onClick={() => handleEdit(task)}
-                          className="text-xs font-medium"
-                        >
-                          <Pencil className="h-3.5 w-3.5 mr-2" /> Edit
-                        </DropdownMenuItem>
-                        <DropdownMenuItem
-                          onClick={() => handleComplete(task)}
-                          className="text-xs font-medium"
-                        >
-                          <CheckCircle2 className="h-3.5 w-3.5 mr-2" /> Complete
-                        </DropdownMenuItem>
-                        <DropdownMenuSeparator />
-                        <DropdownMenuItem
-                          onClick={() => deleteTask(task.id)}
-                          className="text-xs font-medium text-destructive focus:text-destructive"
-                        >
-                          <X className="h-3.5 w-3.5 mr-2" /> Delete
-                        </DropdownMenuItem>
-                      </DropdownMenuContent>
-                    </DropdownMenu>
-                  </motion.div>
-                );
-              })}
-            </div>
-          )}
-
-          {completedTasks.length > 0 && (
-            <div className="space-y-2">
-              <p className="text-[10px] font-black uppercase tracking-widest text-zinc-400 px-1">
-                Completed ({completedTasks.length})
-              </p>
-              {completedTasks.slice(0, 5).map((task, i) => {
-                const typeConfig =
-                  TASK_TYPE_CONFIG[task.task_type] ?? TASK_TYPE_CONFIG.to_do!;
-                if (!typeConfig) return null;
-                const Icon = typeConfig.icon;
-
-                return (
-                  <motion.div
-                    key={task.id}
-                    initial={{ opacity: 0, y: 10 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    transition={{ delay: i * 0.03 }}
-                    className="flex items-start gap-3 p-4 border border-transparent rounded-xl bg-zinc-50/50 group"
-                  >
-                    <motion.div whileTap={{ scale: 0.97 }} className="mt-0.5">
-                      <Checkbox
-                        checked={true}
-                        onCheckedChange={() => handleComplete(task)}
-                        className="h-5 w-5 rounded-md data-[state=checked]:bg-emerald-500 data-[state=checked]:border-emerald-500"
-                      />
-                    </motion.div>
-                    <div
-                      className={cn(
-                        "h-9 w-9 rounded-lg flex items-center justify-center shrink-0 opacity-50",
-                        typeConfig.bgColor,
-                        typeConfig.color,
-                      )}
-                    >
-                      <Icon className="h-4 w-4" />
-                    </div>
-                    <div className="flex-1 min-w-0">
-                      <p className="text-sm font-medium text-zinc-400 line-through">
-                        {task.title}
-                      </p>
-                      {task.completed_at && (
-                        <p className="text-xs text-zinc-400 mt-0.5">
-                          Completed{" "}
-                          {formatDistanceToNow(new Date(task.completed_at), {
-                            addSuffix: true,
-                          })}
-                        </p>
-                      )}
-                    </div>
-                    <Button
-                      variant="ghost"
-                      size="icon"
-                      className="h-8 w-8 rounded-lg opacity-0 group-hover:opacity-100 transition-opacity"
-                      onClick={() => deleteTask(task.id)}
-                    >
-                      <X className="h-4 w-4 text-zinc-400" />
-                    </Button>
-                  </motion.div>
-                );
-              })}
-              {completedTasks.length > 5 && (
-                <p className="text-xs text-zinc-400 text-center py-2">
-                  + {completedTasks.length - 5} more completed tasks
-                </p>
-              )}
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
 
 function useDonorsPageLayout() {
   const { profile, loading: authLoading } = useAuth();
