+++ apps/missionary/app/profile/page.tsx
@@ -69,6 +69,21 @@ import { useState, useCallback, useEffect } from "react";
 import { toast } from "sonner";
 
 import { hasProfileChanges } from "./profile-dirty-state";
+import {
+  asObject,
+  asString,
+  BIO_MAX_CHARS,
+  BIO_MAX_WORDS,
+  BIO_MIN_WORDS,
+  countWords,
+  initialProfile,
+  mapApiProfileToProfileData,
+  PLACEHOLDER_AVATAR,
+  PLACEHOLDER_COVER,
+  TAGLINE_MAX_LENGTH,
+} from "./profile-model";
+
+import type { ProfileData, ValidationErrors } from "./profile-model";
 
 import { PageHeader } from "@/components/page-header";
 import { QuickGive } from "@/features/giving/components/quick-give";
@@ -103,101 +118,6 @@ const gentleTransition = {
   duration: 0.35,
   ease: [0.22, 1, 0.36, 1] as [number, number, number, number],
 };
-
-const TAGLINE_MAX_LENGTH = 100;
-const BIO_MIN_WORDS = 200;
-const BIO_MAX_WORDS = 600;
-const BIO_MAX_CHARS = 3500;
-
-const PLACEHOLDER_AVATAR =
-  "https://images.unsplash.com/photo-1494790108755-2616b612b786?w=400&h=400&fit=crop&crop=face";
-const PLACEHOLDER_COVER =
-  "https://images.unsplash.com/photo-1506905925346-21bda4d32df4?w=1200&h=400&fit=crop";
-
-interface ProfileData {
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  location: string;
-  ministryFocus: string;
-  bio: string;
-  facebook: string;
-  instagram: string;
-  twitter: string;
-  youtube: string;
-  website: string;
-  avatarUrl: string;
-  coverUrl: string;
-}
-
-interface ValidationErrors {
-  firstName?: string;
-  lastName?: string;
-  phone?: string;
-  website?: string;
-  ministryFocus?: string;
-  bio?: string;
-}
-
-const initialProfile: ProfileData = {
-  firstName: "",
-  lastName: "",
-  email: "",
-  phone: "",
-  location: "",
-  ministryFocus: "",
-  bio: "",
-  facebook: "",
-  instagram: "",
-  twitter: "",
-  youtube: "",
-  website: "",
-  avatarUrl: "",
-  coverUrl: "",
-};
-
-type JsonObject = Record<string, unknown>;
-
-function asObject(value: unknown): JsonObject | null {
-  return typeof value === "object" && value !== null
-    ? (value as JsonObject)
-    : null;
-}
-
-function asString(value: unknown): string {
-  return typeof value === "string" ? value : "";
-}
-
-function mapApiProfileToProfileData(apiProfile: unknown): ProfileData {
-  const profile = asObject(apiProfile);
-  const missionary = asObject(profile?.missionary);
-  const socialLinks = asObject(missionary?.social_links);
-
-  return {
-    firstName: asString(profile?.first_name),
-    lastName: asString(profile?.last_name),
-    email: asString(profile?.email),
-    phone: asString(missionary?.phone),
-    location: asString(missionary?.location),
-    ministryFocus: asString(missionary?.tagline),
-    bio: asString(missionary?.bio),
-    facebook: asString(socialLinks?.facebook),
-    instagram: asString(socialLinks?.instagram),
-    twitter: asString(socialLinks?.twitter),
-    youtube: asString(socialLinks?.youtube),
-    website: asString(socialLinks?.website),
-    avatarUrl: asString(profile?.avatar_url),
-    coverUrl: asString(missionary?.cover_url),
-  };
-}
-
-function countWords(text: string): number {
-  return text
-    .trim()
-    .split(/\s+/)
-    .filter((word) => word.length > 0).length;
-}
 
 function ProfileSkeleton() {
   return (
